update user id with int in bigint in DB

+++ src/services/admins/models/models_messages.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Text
+from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Text, BigInteger
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -8,7 +8,7 @@ from src.services.database.database import Base
 class MessageForSending(Base):
     __tablename__ = "message_for_sending"
 
-    user_id = Column(Integer, ForeignKey("users.user_id"), primary_key=True)
+    user_id = Column(BigInteger, ForeignKey("users.user_id"), primary_key=True)
     content = Column(Text, nullable=True)
     photo_path = Column(String(500), nullable=True)
     button_url = Column(String(500), nullable=True)
@@ -19,7 +19,7 @@ class SentMasMessages(Base):
     __tablename__ = "sent_mas_messages"
 
     message_id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.user_id"), nullable=False) # это id админа (напрямую не надо т.к. админ может удалиться)
+    user_id = Column(BigInteger, ForeignKey("users.user_id"), nullable=False) # это id админа (напрямую не надо т.к. админ может удалиться)
     content = Column(Text, nullable=True)
     photo_path = Column(String(500), nullable=True) # фото будем сохранять где-то в архиве
     button_url = Column(String(500), nullable=True)
