+++ src/services/referrals/models/models_ref.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, DateTime
+from sqlalchemy import Column, Integer, ForeignKey, DateTime, BigInteger
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -9,8 +9,8 @@ class Referrals(Base):
     __tablename__ = "referrals"
 
     # Это ID пользователя, которого пригласили (реферала)
-    referral_id = Column(Integer, ForeignKey('users.user_id'), primary_key=True, nullable=False)
-    owner_user_id = Column(Integer, ForeignKey('users.user_id'), nullable=False, index=True) # Это ID пользователя, который пригласил (владельца)
+    referral_id = Column(BigInteger, ForeignKey('users.user_id'), primary_key=True, nullable=False)
+    owner_user_id = Column(BigInteger, ForeignKey('users.user_id'), nullable=False, index=True) # Это ID пользователя, который пригласил (владельца)
     level = Column(Integer, nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
 
@@ -32,9 +32,9 @@ class IncomeFromReferrals(Base):
     replenishment_id = Column(Integer, ForeignKey("replenishments.replenishment_id"), nullable=False)
 
     # ID пользователя-ВЛАДЕЛЬЦА (того, кто получил доход)
-    owner_user_id = Column(Integer, ForeignKey("users.user_id"), nullable=False, index=True)
+    owner_user_id = Column(BigInteger, ForeignKey("users.user_id"), nullable=False, index=True)
     # ID пользователя-РЕФЕРАЛА (с которого был получен доход)
-    referral_id = Column(Integer, ForeignKey("users.user_id"), nullable=False, index=True)
+    referral_id = Column(BigInteger, ForeignKey("users.user_id"), nullable=False, index=True)
 
     amount = Column(Integer, nullable=False)
     percentage_of_replenishment = Column(Integer, nullable=False) # процент от пополнения на момент операции
